base-files: Force changing of the default link if it is wrong
This required due to a side-effect of the way we use multistrap to build the
image. Since it uses upstream debian to do the install, base-files is already
installed from debian, and the link already exists and points to debian. This
change forces the link to the right place if it doesn't match.

Change-Id: I415ff336d5d464bac370e7588cb5babca9953e59